requirements.txt
click
Cython
torch
torchvision
Pillow
timm
# hydra
imagehash
numpy
python-dotenv
transformers
pygame
pytorch-lightning
pydantic
torcheval
tensorboard
# pyaudio
sqlitedict
tensordict
termcolor
torch
torchrl
torchaudio
PyOpenGL
stable-baselines3
gymnasium[all]
gymnasium[atari]
opik
